requirements.txt
customtkinter==5.2.2
mysql-connector-python==8.3.0
Pillow>=10.4.0
reportlab>=4.2.0
python-dotenv==1.0.1